Fix deprecation warning when building sleepers data frame
When a player on the sleepers list was drafted, they are filter from the main player data.
They way we were building the sleepers data was causing the player to be readded as a player with no data.
By switching to the dataframe `.filter()`, the player is no longer readded and we no longer get the deprecation warning.